Make humanized timespans optional
Switching every Azure Pipelines user to humanized durations was not
wanted, so the format now sits behind BuildSystems_HumanizedTimespans,
off by default and following the BuildSystems_FailOnFatal schema.

The humanizer moves to Cake.Module.Shared, and FormatTime and
FormatDuration read the setting themselves. That costs them their
static modifier, since the setting comes from the Cake context, but it
keeps the flag out of every call site.
