+++ src/Cake.Module.Shared/Extensions.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Globalization;
+
 using Cake.Core.Configuration;
 
 namespace Cake.Module.Shared
@@ -7,6 +10,30 @@ namespace Cake.Module.Shared
     /// </summary>
     public static class Extensions
     {
+        private static readonly string SubSecondDuration = "< 1sec";
+
+        /// <summary>
+        /// Renders a <see cref="TimeSpan"/> the way a person reads a stopwatch, as <c>m:ss</c>, or as
+        /// <c>h:mm:ss</c> from an hour upwards. Anything below a second becomes <c>&lt; 1sec</c>.
+        /// </summary>
+        /// <param name="time">The <see cref="TimeSpan"/> to render.</param>
+        /// <returns>The humanized duration.</returns>
+        public static string Humanize(this TimeSpan time)
+        {
+            // A tenth of a second means nothing in a build summary, but the difference between "it ran
+            // instantly" and "it took a second" does, so anything below a second is called out as such.
+            if (time < TimeSpan.FromSeconds(1))
+            {
+                return SubSecondDuration;
+            }
+
+            var rounded = TimeSpan.FromSeconds(Math.Round(time.TotalSeconds, MidpointRounding.AwayFromZero));
+
+            return rounded < TimeSpan.FromHours(1)
+                ? string.Format(CultureInfo.InvariantCulture, "{0}:{1:00}", (long)rounded.TotalMinutes, rounded.Seconds)
+                : string.Format(CultureInfo.InvariantCulture, "{0}:{1:00}:{2:00}", (long)rounded.TotalHours, rounded.Minutes, rounded.Seconds);
+        }
+
         /// <summary>
         /// Get a config-value as a flag from the <see cref="ICakeConfiguration"/>.
         /// </summary>
